validator-cdk/requirements.txt
-e .
aws-cdk.aws-lambda==1.93.0
aws-cdk.aws_apigateway==1.93.0
aws-cdk.aws-apigatewayv2-integrations==1.93.0

